Self trial, adding a faction - Taris
As a trial to see how things work, I added Taris as a playable faction.
check it out and tell me what you think. It could use some polishing,
such as Taris’ first moon being too close, and the Rogue Moon should be
close to Paris, and in an Asteroid Belt.