Yavin Update
An update for Yavin. I also made a “full” version of it and placed it
in a separate category, just in case we ever need or wanted to come
back and use it in the future. A couple of edits to Taris system as
well, like giving it a G class Star.